never instantiate specific Query in tests directly
